Fixed autoload, added support for Bundle ether port channels
